process ncer better

+++ module-api/src/main/java/com/szadowsz/nds4j/data/nfs/cells/CellPojo.java
@@ -9,14 +9,14 @@ public class CellPojo {
         short minX;
         short minY;
 
-        short[][] attrs;
+        int[][] attrs;
 
         public CellPojo(int cellCount, int cellAttr) {
                 this.nAttribs = cellCount;
                 this.cellAttr = cellAttr;
-                this.attrs = new short[nAttribs][];
+                this.attrs = new int[nAttribs][];
                 for (int i = 0; i < this.nAttribs; i++) {
-                        this.attrs[i] = new short[3];
+                        this.attrs[i] = new int[3];
                 }
         }
 
@@ -29,18 +29,18 @@ public class CellPojo {
 
         public void setEmptyAttributes() {
                 nAttribs = 1;
-                this.attrs = new short[nAttribs][];
-                attrs[0] = new short[3];
+                this.attrs = new int[nAttribs][];
+                attrs[0] = new int[3];
                 attrs[0][0] = 0x0200; // Disable rendering
         }
 
-        public void setOamAttrs(int index, short attr0, short attr1, short attr2) {
+        public void setOamAttrs(int index, int attr0, int attr1, int attr2) {
                 attrs[index][0] = attr0;
                 attrs[index][1] = attr1;
                 attrs[index][2] = attr2;
         }
 
-        public short[] getOamAttrs(int index) {
+        public int[] getOamAttrs(int index) {
                 return attrs[index];
         }
 }
